Move da cost calc stuff to its own mod
